rainnet/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
captum==0.8.0
certifi==2025.8.3
cftime==1.6.4.post1
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cmweather==0.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dask==2025.9.1
Deprecated==1.2.18
deprecation==2.1.0
eval_type_backport==0.2.2
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.75.0
h5netcdf==1.6.4
h5py==3.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lat_lon_parser==1.3.1
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
meteors==0.2.0
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
netCDF4==1.7.2
networkx==3.5
numpy==1.26.4
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
partd==1.4.2
pillow==11.3.0
protobuf==5.29.5
pyarrow==21.0.0
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pyproj==3.7.2
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
spyndex==0.7.1
sympy==1.14.0
termcolor==3.1.0
tf-keras-vis==0.8.7
threadpoolctl==3.6.0
tifffile==2025.10.4
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
ttach==0.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wradlib==2.4.0
wrapt==1.17.3
xarray==2025.9.0
xmltodict==1.0.2
xradar==0.10.0
